refactor(frontend): move copying hotkeys into CopyHotkeys component
Closes #188
